+++ R/NMorderColumns.R
@@ -12,7 +12,7 @@
 ##' @param lower.last Should columns which names contain lowercase
 ##'     characters be moved towards the back? Some people use a
 ##'     standard of lowercase variables (say "race") being character
-##'     representations ("asian", "caucasian", etc.) variables and the
+##'     representations ("Asian", "Caucasian", etc.) variables and the
 ##'     uppercase (1,2,...) being the numeric representation for
 ##'     Nonmem.
 ##' @param chars.last Should columns which cannot be converted to