+++ NEWS.md
@@ -1,10 +1,67 @@
 # Since NMdata 0.0.11
+Thanks to Bill Denney for discussions and inputs to this release!
+
+* New function: cl - creates factors, ordered by the appearance of the
+  elements when created. cl("b","a") results in a factor with levels
+  "b" and "a". This can save quite some typing in data set
+  preparation.
+
+* New function: fnAppend - append a string to a file name before the
+  file name extension. fnAppend("data.csv","subset") results in
+  "data_subset.csv".
+
+* General support for a file.data argument when a specific input data
+  file is to be used instead of finding this information in the
+  control streams. This is very useful if you archive input data
+  together with a nonmem run in a way that the path in the control
+  stream has to be overruled. Like many other of this type of
+  arguments in NMdata, it can be a function that systematically
+  converts the path to the control stream to the input data
+  archive. Running Nonmem this way breaks the link between an input
+  dataset that may change over time and the model runs that become
+  self-contained packages of input and output data.
+
+* Support for NOHEADER option in Nonmem $TABLE blocks. If NMdata is
+  used to read the results, there is no need to use NOHEADER (which
+  opens the door to mistakes if manually renaming the columns in
+  results), but NMdata should now also be able to handle this. 
+
+* If found in data, CMT is added to the breakdown of rows when
+  summarizing results from NMscanData. Before, it was broken down on
+  EVID only.
+
+* Support for non-event (say for $PRED) datasets in NMcheckData.
+
+* Support for custom column names for DV (col.dv) and MDV (col.mdv) in
+  NMcheckData.
+
+* Support for file.mod and dir.data arguments in NMcheckData when
+  running on a control stream.
+
+* NMgenText now has an argument called until that specifies the last
+  column(s) to include in $INPUT.
+
+* If NMgenText does not find any Nonmem-compatible columns to report,
+  it now throws a warning and returns NULL.
+
 ## Bugfixes
 * NMextractDataFile was not cleaning all paths correctly. This mostly
   impacts NMcheckData's ability to find data files when using the file
   argument in NMcheckData. Only affects certain models.
 * NMextractDataFile was not working with dir.data. Fixed.
 * NMextractDataFile Now handles absolute paths correctly.
+
+## Minor changes 
+* compareCols now by default lists the columns where no differences
+  were found. 
+
+* NMreadTab throws a message instead of a warning in case duplicate
+  column names are found and removed.
+	  
+* NMwriteData now runs NMgenText in try, just in case.
+
+* fnExtension now supports adding extensions to strings without
+  extensions, ie. fnExtension("file",".txt").
 
 # NMdata 0.0.11
 
@@ -56,7 +113,7 @@
   for the corresponding rds files. The function is not new in NMdata
   but was not exported until 0.0.10.
 
-* cc is a function that creates character vectors from argmuments
+* cc is a function that creates character vectors from arguments
   without quotes. This is just to be able to skip typing quotes when
   listing say column names. So do cc(a,b,c) to get the exact same as
   c("a","b","c"). You cannot do this with strings that contain special
@@ -78,7 +135,7 @@
 
 * flagsAssign has got a few updates related to separate handling of
   different types of events. Often, this will be used to assign flags
-  to observations, doses etc. separately. You can easily speficy a
+  to observations, doses etc. separately. You can easily specify a
   subset of data to run flagsAssign on, and it will by default check
   for whether values of EVID are unique. This is similar to what
   flagsCount does.
@@ -123,7 +180,7 @@ that tests pass after the release of data.table v1.14.2.
   write meta data. Meta data is stored as an attribute to the data
   object (attributes(data)$NMdata).
 
-* Translation table includes a column ranking the match betwen input
+* Translation table includes a column ranking the match between input
   data file contents and $INPUT. OK: names match, diff: names do not
   match, off: diff and name is found elsewhere.
 
@@ -257,7 +314,7 @@ number of columns added to df1 against expectation.
 
 egdt is a new function for expanding grids of data.tables. This is
 quite technical, and it fills a whole when constructing data with
-data.tables. It mimicks the behavior of merge.data.frame on objects
+data.tables. It mimics the behavior of merge.data.frame on objects
 with no common columns.
 
 ## Bugfixes related to 
@@ -276,7 +333,7 @@ names.
 
 The exclusion flag functions flagsAssign and flagsCount have been
 generalized to use customizable column names for the numerical and
-caracter flags. The default can be configured using NMdataConf.
+character flags. The default can be configured using NMdataConf.
 
 # NMdata 0.0.6.4
 If all common column names two data objects to merge are not used for
@@ -293,7 +350,7 @@ Meta information added about the input data.
 
 # NMdata 0.0.6.2
 A summary function is provided for NMdata objects. There is a print
-function for the summary too. This is printet automatically by NMdata
+function for the summary too. This is printed automatically by NMdata
 unless quiet=TRUE.
 
 A lot of meta information has been added in an attribute to NMdata
@@ -320,7 +377,7 @@ used for the individual functions too.
 The translation from the output control stream file path (.lst in PSN)
 and the input control stream (.mod in PSN) can now be configured
 through the option "NM.file.mod". Typically, all the models to be
-considered in an anlysis have been run on the same system, so it makes
+considered in an analysis have been run on the same system, so it makes
 most sense to define this behavior once and for all for most users.
 
 NMwriteData improved with checks of column names and automated
@@ -355,7 +412,7 @@ now treated like FIRSTONLY while FIRSTLASTONLY tables are disregarded
 The most obvious change since 0.0.3 is that only one data.table is
 being returned from NMscanData. This is what used to be the `row`
 element in the returned objects previously. The main reason for this
-change is that it makes it easier for users to postprocess only one
+change is that it makes it easier for users to post-process only one
 dataset before splitting into different levels of variability. The
 small cost is that the user will have to run findCovs, or findVars to
 get the desired level of variability. These functions are however very